Ditch old code in favor or serverless infra
I don't want to run a server for this hobby project. So, in an attempt
to reduce cost and complexity, use the AWS Lambda serverless
infrastructure, and delete the old code that used Slack Bolt SDK.

I am also trying to avoid installing external dependencies, so ditch
the super handy `requests` library in favor of the lower level, `urllib`
python standard library. This makes the code uglier but my life easier.